Add `getWordlists()`
This adds a get Word lists functions that returns all the wordlists from the wordlist database.  This may need to be changed to context pack? I am not sure what would be more useful in the end, but I have it listed as wordlist as of right now.  This also adds the necessary classes (Wordlist class and Word class)