Add sign in and sign out buttons
You can now sign in with your google account but only if I've given you access to sign in since google likes security while in development. You can go into the console and see your account information and the id token that will need to eventually be sent to the server so it can identify what kind of user you are. The signout button signs you out but doesn't do anything else.