Update server urls
This updates the post urls in the server file.  This makes it so the word list edits and the contextpack edits have their own url.  I did this because these functions are currently separate operations, and I think it makes more sense to call them, and organize them as such based on the current code.