editing wordlists
This adds tests for editing the main attributes of a word list (the name and the enabled attributes) I would like to add icon as well once we figure out what we want the icons to look like (We are still working through this). This tests that a post request made correctly will delete a noun by word.