Add word logic
This adds logic to the edit word lists method that allows the user to add words and their forms to a word list. I am just now thinking that we may want to add a some sort of validator that will only add the word if it isnt already part of the wordlist?