Alter display words
This changes the display word functin and the way these words are displayed. KK expressed that she liked the display view being reduced to just the word without its forms, so this does that.  it also displays the words as mat-chips, rather than as a paragraph, I believe that this is a cleaner look and more closely reflects KK's wants from the last showcase.