Set client post methods to match server
This matches up some of the client side methods with the server side post methods.  The functions for communication with the server are still not working, and I have not been able to figure that out yet, although I do believe it is something client side because I am able to make the requests through Postman just fine.