Add deleting nouns
This adds a very ugly version of deleting nouns.  I don't have time to clean it up right now, but will come back later.  Nouns can be deleted in 'editing mode' which at the moment is triggered by pressing the 'plus' fab (not a very good choice for an icon I agree)