Fix null values for adding words with no forms
There was an issue with the addWord() function, where null was being added as a form (and displaying as a form) if the user doesn't add their own forms to the word.  This trims off the null values so that null is not added as a form for these words.