Color change and moving remove word button
I moved the remove word button to be on the same line as the input for the word. Also change the colors in the context pack-card to match the colors I picked for the add context pack. I also moved the placement of verbs above adjectives to match the layout of the words in the contedt pack cards.