Add delete wordlist HTML
This adds HTML for removing a wordlist from a context pack.  HTML wise I only made this action available when in editing mode (when the pencil button has been clicked), and only for admins.  I did this because only admins can edit wordlists right now, so having the HTML to delete the wordlist, but not being able to actually delete would be misleading.