Create a Learner card
This creates a simple learner card that as of now just shows the learner's name, and which contextpacks they have been assigned.  it is a big ugly right now but this is just a start, so we can get a better idea as to what we want to do for design.