Touch up assigned Words
KK expressed that should would like the words to be alphabetized, and that she would like the wordlist name where each word originated from to pop up on hover.  This commit makes those changes. The group also decided that the colors representing each part of speech should be different.  That change is made here as well.