Display Assigned Wordlists
This adds the learners specific word lists to the list under their respective context packs.  This adds a couple of helper functions which help achieve that.  This includes a helper function which checks to see if a specific word list has been disabled for a learner, if that word list is disabled, this word list will not be shown