Test for assigning a wordlist
Since a learner has a list of disabled wordlists, assigning a wordlist will just be removing the wordlist from the disabled wordlist list.  We set it up this way, because we are assuming a wordlist to be enabled for a learner by default when they are assigned a context pack.