assign list when checked
This adds the logic for a assigning word lists to learners by checking the box next to the wordlist in the drop down menu.  This includes updating the assigned words, assigned wordlists.  As of right now, un-checking a box does not do anything, the logic for un-assigning a wordlist does not yet exist.