Edit tests for contextpack-list.spec
Adds a wait time to fix a timing issue to the viewInfo test, the test now passes.
Specifies to look at only the first delete verb button. The test still does not pass, and the issue might be that (be.visible) is the wrong test to use.
Replaces the addNoun test with what was there initially, but I haven't gotten the test to pass again.
Co-authored-by: Isabelle Hjelden <hjeld013@morris.umn.edu>