Update learner-list.spec tests to test logged in user view
Updating the html so that only logged in users can view learners broke some of the e2e tests. This commit updates them to test that users who aren't logged in cannot see any learners, and users who are logged in are able to see them and search for them.
It also tests for the proper title ("My learners"), and tests for the proper message ("Please sign in to view learners"). All tests pass