Add tools section
This adds a section to the readme that briefly covers the major tools and provides links to the documentation/ main sites for this tools.  This includes testing tools, workflow tools, languages, etc.  This list may not be completely exhaustive and others are welcome to add to it should they see fit